Move speed-to-distance conversion to vehicle_controller
- Relocate conversion logic from webots_controller to vehicle_controller
- Change message type to Float64MultiArray to support multiple tire values
- Update tire radius to 0.374m (from BmwX5Wheel.proto)
- Improves modularity by centralizing conversion logic
